+++ packages/delegation-deployments/src/index.ts
@@ -31,6 +31,7 @@ export const CHAIN_ID = {
   mantleMainnet: 0x1388,
   katanaMainnet: 0xb67d2,
   intuitionMainnet: 0x483,
+  robinhoodMainnet: 0x1237,
   // Testnets
   bscTestnet: 0x61,
   arbitrumSepolia: 0x66eee,
@@ -55,6 +56,7 @@ export const CHAIN_ID = {
   mantleSepolia: 0x138b,
   katanaBokuto: 0xb405d,
   intuitionTestnet: 0x350b,
+  robinhoodTestnet: 0xb626,
   // decommissioned
   lineaGoerli: 0xe704,
 };
@@ -122,6 +124,7 @@ export const DELEGATOR_CONTRACTS: DeployedContracts = {
     [CHAIN_ID.mantleMainnet]: DEPLOYMENTS_1_3_0,
     [CHAIN_ID.katanaMainnet]: DEPLOYMENTS_1_3_0,
     [CHAIN_ID.intuitionMainnet]: DEPLOYMENTS_1_3_0,
+    [CHAIN_ID.robinhoodMainnet]: DEPLOYMENTS_1_3_0,
     // Testnets
     [CHAIN_ID.bscTestnet]: DEPLOYMENTS_1_3_0,
     [CHAIN_ID.citreaTestnet]: DEPLOYMENTS_1_3_0,
@@ -146,5 +149,6 @@ export const DELEGATOR_CONTRACTS: DeployedContracts = {
     [CHAIN_ID.mantleSepolia]: DEPLOYMENTS_1_3_0,
     [CHAIN_ID.katanaBokuto]: DEPLOYMENTS_1_3_0,
     [CHAIN_ID.intuitionTestnet]: DEPLOYMENTS_1_3_0,
+    [CHAIN_ID.robinhoodTestnet]: DEPLOYMENTS_1_3_0,
   },
 };